requirements.txt
absl-py==1.1.0
aiofiles==0.8.0
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
altgraph==0.17.2
anyio==3.5.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.1
atomicwrites==1.4.0
attrs==21.2.0
autobahn==22.1.1
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
barnum==0.5.1
beautifulsoup4==4.9.3
blis==0.7.5
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.6
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.4
ci-info==0.2.0
click==8.0.3
cloudinary==1.29.0
colorama==0.4.4
comtypes==1.1.11
configobj==5.0.6
configparser==5.2.0
constantly==15.1.0
countryinfo==0.1.2
cryptography==36.0.1
cssselect==1.1.0
cssselect2==0.6.0
cssutils==2.4.2
cv==1.0.0
cx-Freeze==6.9
cx-Logging==3.0
cycler==0.11.0
cymem==2.0.6
daphne==3.0.2
decorator==4.4.2
dj-database-url==0.5.0
Django==4.0.2
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-ckeditor==6.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-cropper-image==1.0.5
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-livereload-server==0.4
django-mathfilters==1.0.0
django-mysql==4.5.0
django-online-users==0.3
django-redis-cache==3.0.0
djangorestframework==3.13.1
easy-thumbnails==2.8.1
et-xmlfile==1.1.0
etelemetry==0.3.0
fake-useragent==0.1.11
Faker==13.11.1
fastapi==0.72.0
fb-down==1.0.1
filelock==3.0.12
filetype==1.0.9
fitz==0.0.1.dev2
flatbuffers==2.0
fonttools==4.31.2
frontend==0.0.3
frozenlist==1.3.0
future==0.18.2
gdown==4.4.0
geosky==1.0.9
googletrans==3.1.0a0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
heroku==0.1.4
hiredis==2.0.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
huepy==1.2.1
huggingface-hub==0.0.16
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.10.0
incremental==21.3.0
iniconfig==1.1.1
insta-scrape==2.1.2
instabot==0.117.0
instaloader==4.9.2
isodate==0.6.1
itemadapter==0.5.0
itemloaders==1.0.4
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.0.1
jotform==1.2
kiwisolver==1.4.2
langcodes==3.3.0
llvmlite==0.38.0
loader==2017.9.11
Logbook==1.5.3
lxml==4.8.0
macholib==1.16
MarkupSafe==2.0.1
matplotlib==3.5.1
mediapipe==0.8.10
mhyt==3.5.5
mock==4.0.3
modulegraph==0.19.2
MouseInfo==0.1.3
moviepy==1.0.3
mozscape==0.1.1
msgpack==1.0.3
multidict==5.1.0
mupdf==1.19.0.20210929.1226
murmurhash==1.0.6
networkx==2.6.3
nibabel==3.2.2
nipype==1.7.0
nltk==3.6.2
numba==0.55.1
numpy==1.21.5
onnxruntime==1.10.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.3.56
opencv-python-headless==4.5.2.52
openpyxl==2.6.3
outcome==1.1.0
packaging==21.0
parrot==0.0.16
parsel==1.6.0
pathlib==1.0.1
pathy==0.6.1
pdfminer.six==20211012
pdfplumber==0.6.0
pdfrw==0.4
pefile==2021.9.3
pendulum==2.1.2
Pillow==9.0.1
pluggy==1.0.0
premailer==3.10.0
preshed==3.0.6
proglog==0.1.9
promptlib==1.0.0
Protego==0.2.1
protobuf==3.20.1
prov==2.0.0
psutil==5.9.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.11.0
py2app==0.28
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycaw==20220416
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
PyDispatcher==2.0.5
pydot==1.4.2
PyGetWindow==0.0.9
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyJWT==2.3.0
PyMatting==1.1.5
PyMsgBox==1.0.9
PyMuPDF==1.19.5
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqt5-tools==5.15.4.3.2
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
PyRect==0.1.4
PyScreeze==0.1.27
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-magic==0.4.25
python-multipart==0.0.5
pytube==11.0.1
PyTweening==1.0.3
pytz==2021.1
pytzdata==2020.1
PyWavelets==1.3.0
pywhatkit==5.1
pywin32==303
pywin32-ctypes==0.2.0
pyxnat==1.4
PyYAML==5.4.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
queuelib==1.6.2
rdflib==6.1.1
redis==3.5.3
redmail==0.4.0
regex==2021.8.28
rembg==2.0.13
reportlab==3.6.10
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
responses==0.20.0
rfc3986==1.5.0
rq==1.9.0
rsa==4.8
sacremoses==0.0.45
schedule==1.1.0
scikit-image==0.19.1
scikit-learn==0.24.2
scipy==1.7.3
scrapetube==2.2.2
Scrapy==2.6.1
selenium==4.1.3
sentence-splitter==1.4
sentence-transformers==2.0.0
sentencepiece==0.1.96
seolib==0.1.3
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
srsly==2.4.2
starlette==0.17.1
stripe==2.64.0
svglib==1.3.0
Telethon==1.24.0
termcc==2019.8.2
terminaltables==3.1.0
thinc==8.0.13
threadpoolctl==2.2.0
tifffile==2022.5.4
tinycss2==1.1.1
tldextract==3.3.0
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.1
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.3
traits==6.3.2
transformers==4.10.2
trio==0.20.0
trio-websocket==0.9.2
twilio==7.6.0
Twisted==22.1.0
twisted-iocpsupport==1.0.2
twocheckout==0.4.0
txaio==21.2.1
typer==0.4.0
typing-extensions==3.10.0.0
tzdata==2021.5
urllib3==1.26.6
utils==1.0.1
uvicorn==0.17.0
w3lib==1.22.0
Wand==0.6.7
wasabi==0.9.0
watchdog==2.1.7
webdriver-manager==3.5.4
webencodings==0.5.1
whitenoise==5.3.0
wikipedia==1.4.0
worker==2018.4.1
wsproto==1.1.0
xlrd==1.2.0
yagmail==0.15.277
yarl==1.6.3
youtube-dl==2021.12.17
zipp==3.6.0
zope.interface==5.4.0
